requirements.txt
absl-py==2.4.0
alphashape==1.3.1
cbor2==5.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-log==0.4.0
coacd==1.0.7
etils==1.14.0
fast_simplification==0.1.13
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
idna==3.11
imgui-bundle==1.92.601
markdown-it-py==4.0.0
mdurl==0.1.2
meshio==5.3.5
mujoco==3.5.0
mujoco-warp==3.5.0.2
networkx==3.6.1
newton==1.0.0
newton-actuators==0.1.0
newton-usd-schemas==0.1.0
numpy==2.4.4
pillow==12.1.1
pycollada==0.9.3
pyglet==2.1.13
Pygments==2.20.0
PyOpenGL==3.1.10
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
resolve-robotics-uri-py==0.4.1
rich==14.3.3
rtree==1.4.1
scipy==1.17.1
shapely==2.1.2
six==1.17.0
smmap==5.0.3
trimesh==4.11.5
typing_extensions==4.15.0
urllib3==2.6.3
usd-core==26.3
warp-lang==1.12.0
zipp==3.23.0
